restar mitad de puntos al perder

+++ Castigo Divino/Assets/Scripts/Magnet.cs
@@ -8,7 +8,7 @@ public class Magnet : MonoBehaviour
     public float magnetStrength = 5f;  
     public float magnetRange = 5f;    
 
-    [SerializeField] private float amountPoints;
+    [SerializeField] private int amountPoints;
     [SerializeField] private PointsUI pointsUI;
     [SerializeField] private Loot loot;
